moves the health providing caching agent from aws to core
