+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/controllers/AmazonNamedImageLookupController.groovy
@@ -30,8 +30,8 @@ import org.springframework.web.bind.annotation.RequestMethod
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 
-import static com.netflix.spinnaker.clouddriver.aws.data.Keys.Namespace.IMAGES
-import static com.netflix.spinnaker.clouddriver.aws.data.Keys.Namespace.NAMED_IMAGES
+import static com.netflix.spinnaker.clouddriver.core.provider.agent.Namespace.IMAGES
+import static com.netflix.spinnaker.clouddriver.core.provider.agent.Namespace.NAMED_IMAGES
 
 @Slf4j
 @RestController
