+++ clouddriver-titus/clouddriver-titus.gradle
@@ -1,6 +1,5 @@
 dependencies {
   compile project(":clouddriver-core")
-  compile project(":clouddriver-aws")
   compile spinnaker.dependency('bootActuator')
   compile spinnaker.dependency('bootWeb')
   compile 'com.squareup.retrofit2:retrofit:2.0.0-beta4'
